refactor: rename workers to single-letter names (N, O, S, Z)
- nuclei-signal -> N (Nuclei signal worker)
- openhack-hunter -> O (OpenHack hunter worker)
- strix-core -> S (Strix core worker)
- zap-signal -> Z (ZAP signal worker)

Updated all references in:
- Worker directories and subdirectories (cmd/, env/)
- go.mod module names
- Dockerfile build paths and binary names
- main.go service names
- .env.example files
- go.work file
- infra/docker-compose/workers.yml
