+++ workers/N/go.mod
@@ -1,0 +1,3 @@
+module openhunter/workers/N
+
+go 1.22
