+++ workers/O/go.mod
@@ -1,0 +1,3 @@
+module openhunter/workers/O
+
+go 1.22
