+++ workers/S/go.mod
@@ -1,0 +1,3 @@
+module openhunter/workers/S
+
+go 1.22
