+++ workers/Z/go.mod
@@ -1,0 +1,3 @@
+module openhunter/workers/Z
+
+go 1.22
