--- workers/nuclei-signal/go.mod
@@ -1,3 +1,0 @@
-module openhunter/workers/nuclei-signal
-
-go 1.22
