--- workers/openhack-hunter/go.mod
@@ -1,3 +1,0 @@
-module openhunter/workers/openhack-hunter
-
-go 1.22
