--- workers/strix-core/go.mod
@@ -1,3 +1,0 @@
-module openhunter/workers/strix-core
-
-go 1.22
