--- workers/zap-signal/go.mod
@@ -1,3 +1,0 @@
-module openhunter/workers/zap-signal
-
-go 1.22
